--- debug-utility/neoforge/src/main/resources/META-INF/neoforge.mods.toml
@@ -1,16 +1,0 @@
-modLoader = "javafml"
-loaderVersion = "[1,)"
-license = "All rights reserved"
-[[mods]]
-modId = "debug_utils"
-version = "${file.jarVersion}"
-displayName = "DebugUtils"
-credits = "Kotori316"
-authors = "Kotori316"
-description = ''' '''
-[[dependencies.debug_utils]]
-modId = "neoforge"
-type = "required"
-versionRange = "[21, )"
-ordering = "NONE"
-side = "BOTH"
