--- debug-utility/neoforge/src/main/resources/pack.mcmeta
@@ -1,6 +1,0 @@
-{
-  "pack": {
-    "description": "Debug Utility",
-    "pack_format": 18
-  }
-}
