+++ neoforge/src/main/java/com/kotori316/testutil/TestUtilMod.java
@@ -3,16 +3,14 @@ package com.kotori316.testutil;
 import com.kotori316.testutil.common.TestUtilityCommon;
 import com.kotori316.testutil.common.reporter.ReporterRegister;
 import net.neoforged.bus.api.IEventBus;
-import net.neoforged.fml.common.Mod;
 import net.neoforged.neoforge.event.RegisterGameTestsEvent;
 
-@Mod(TestUtilityCommon.MOD_ID)
 public final class TestUtilMod {
-    public TestUtilMod(IEventBus modBus) {
-        modBus.addListener(this::changeReporter);
+    public static void register(IEventBus modBus) {
+        modBus.addListener(TestUtilMod::changeReporter);
     }
 
-    void changeReporter(RegisterGameTestsEvent event) {
+    static void changeReporter(RegisterGameTestsEvent event) {
         ReporterRegister.changeReporter();
         TestUtilityCommon.GENERAL.info("Inject CreateFileReporter by {}", TestUtilityCommon.MOD_ID);
     }
