+++ neoforge/src/main/java/com/kotori316/testutil/mixin/MixinHashCache.java
