+++ neoforge/src/test/java/com/kotori316/debug/neoforge/DebugUtilsNeoForgeTest.java
