+++ neoforge/src/test/java/com/kotori316/testutil/MCTestInitializerTest.java
